+++ tracklytics-runtime/src/test/java/com/orhanobut/tracklytics/debugger/EventQueueTest.java
@@ -7,7 +7,7 @@ import org.junit.Test;
 import java.util.Collections;
 import java.util.List;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static com.google.common.truth.Truth.assertThat;
 import static org.mockito.Matchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;